requirements.txt
allure-pytest-bdd==2.14.2
selenium==4.33.0